RavenDB-18880: abstract subscription state & storage
