+++ src/Raven.Server/Documents/Subscriptions/Sharding/ShardSubscriptionStorage.cs
@@ -1,0 +1,56 @@
+﻿using Raven.Client.Documents.Subscriptions;
+using Raven.Client.Exceptions.Documents.Subscriptions;
+using Raven.Client.Json.Serialization;
+using Raven.Client.ServerWide;
+using Raven.Server.Documents.Sharding;
+using Raven.Server.ServerWide;
+using Raven.Server.ServerWide.Context;
+
+namespace Raven.Server.Documents.Subscriptions.Sharding;
+
+public class ShardSubscriptionStorage : SubscriptionStorage
+{
+    public ShardSubscriptionStorage(ShardedDocumentDatabase db, ServerStore serverStore) : base(db, serverStore)
+    {
+    }
+
+    public override void HandleDatabaseRecordChange(DatabaseRecord databaseRecord)
+    {
+        using (_serverStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
+        using (context.OpenReadTransaction())
+        {
+            //checks which subscriptions should be dropped because of the database record change
+            foreach ((long id, var state) in _subscriptions)
+            {
+                var subscriptionName = state.SubscriptionName;
+                if (subscriptionName == null)
+                    continue;
+
+                using var taskStateRaw = _serverStore.Cluster.Subscriptions.ReadSubscriptionStateRaw(context, _databaseName, subscriptionName);
+                if (taskStateRaw == null)
+                {
+                    // the subscription is deleted
+                    continue;
+                }
+
+                var taskState = JsonDeserializationClient.SubscriptionState(taskStateRaw);
+                var whoseTaskIsIt = GetSubscriptionResponsibleNode(databaseRecord, taskState);
+                if (whoseTaskIsIt != _serverStore.NodeTag)
+                {
+                    DropSubscriptionConnections(id,
+                        new SubscriptionDoesNotBelongToNodeException($"Shard subscription '{id}' operation was stopped, because it's now under a different server's responsibility"));
+                }
+            }
+        }
+    }
+
+    protected override string GetSubscriptionResponsibleNode(DatabaseRecord databaseRecord, SubscriptionState taskStatus)
+    {
+        return _serverStore.WhoseTaskIsIt(databaseRecord.Topology, taskStatus, taskStatus);
+    }
+
+    internal override void CleanupSubscriptions()
+    {
+        // all shard workers are disposed by orchestrator directly
+    }
+}
