+++ src/Raven.Server/ServerWide/Commands/Subscriptions/PutSubscriptionCommand.cs
@@ -135,7 +135,7 @@ namespace Raven.Server.ServerWide.Commands.Subscriptions
         protected void RemoveSubscriptionStateFromStorage(ClusterOperationContext context, long subscriptionId)
         {
             var subscriptionStateTable = context.Transaction.InnerTransaction.OpenTable(ClusterStateMachine.SubscriptionStateSchema, ClusterStateMachine.SubscriptionState);
-            using (SubscriptionConnectionsStateBase.GetDatabaseAndSubscriptionPrefix(context, DatabaseName, subscriptionId, out var prefix))
+            using (AbstractSubscriptionConnectionsStateBase.GetDatabaseAndSubscriptionPrefix(context, DatabaseName, subscriptionId, out var prefix))
             {
                 using var _ = Slice.External(context.Allocator, prefix, out var prefixSlice);
                 subscriptionStateTable.DeleteByPrimaryKeyPrefix(prefixSlice);
